enable cors on Quarkus tests
This will not play nicely with frontend-starter-app, or (I'm assuming)
any of the other insights frontend apps. insights-chrome places a
Content Security Policy that has a set of defined sources we can fetch
from.

Instead, I'm thinking of writing the framework for the UI in this repo,
and then copy-pasting it over into a patternfly-seed app, where I can
reliably hit this fake rest service.